Move event releated depencies to EventServiceProvider
